+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/organization/OrganizationDataSourceImpl.kt
@@ -1,5 +1,6 @@
-package com.daedan.festabook.data.datasource.remote
+package com.daedan.festabook.data.datasource.remote.organization
 
+import com.daedan.festabook.data.datasource.remote.ApiResult
 import com.daedan.festabook.data.model.response.OrganizationResponse
 import com.daedan.festabook.data.service.OrganizationService
